requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.6
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==36.2.2
flake8==7.1.2
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
phonenumbers==8.13.55
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
